requirements.txt
fastapi==0.115.13
litserve==0.2.12
numpy==2.3.1
onnxruntime==1.22.0
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
python_jose==3.5.0
Requests==2.32.4
scipy==1.16.0
SQLAlchemy==1.4.54
streamlit==1.46.0
transformers==4.51.3
